update location of defaultPropTestParams to not accidentally pull in jest
