+++ test/property/default_prop_test_params.ts
@@ -1,0 +1,29 @@
+import * as fc from 'fast-check';
+
+export function defaultPropTestParams<T = unknown>(): fc.Parameters<T> {
+    const baseParams = {
+        numRuns: Number(process.env.AZLE_PROPTEST_NUM_RUNS ?? 1),
+        reporter: (runDetails: fc.RunDetails<T>): void => {
+            const seed = runDetails.seed;
+            const path = runDetails.counterexamplePath;
+            const reproductionCommand = `AZLE_PROPTEST_SEED=${seed}${path !== null ? ` AZLE_PROPTEST_PATH="${path}"` : ''} AZLE_VERBOSE=true AZLE_TEMPLATE=true npm test`;
+            const reproductionMessage = `To reproduce this exact test case, run:\n${reproductionCommand}`;
+            console.info(reproductionMessage);
+            if (runDetails.failed) {
+                throw new Error(
+                    `${reproductionMessage}\n\n${fc.defaultReportMessage(runDetails)}`
+                );
+            }
+        },
+        endOnFailure: process.env.AZLE_PROPTEST_SHRINK === 'true' ? false : true
+    };
+
+    const seed =
+        process.env.AZLE_PROPTEST_SEED !== undefined
+            ? Number(process.env.AZLE_PROPTEST_SEED)
+            : undefined;
+
+    const path = process.env.AZLE_PROPTEST_PATH;
+
+    return seed !== undefined ? { ...baseParams, seed, path } : baseParams;
+}
